ai-server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
google-generativeai==0.7.2
pydantic==2.9.2
python-dotenv==1.0.1